requirements.txt
django-model-utils>=2.0
django-environ>=0.4.5
psycopg2-binary>=2.8.4
django-extensions>=2.2.5
django-easy-audit>=1.1.1
django-model-utils>=2.0
django-adminplus>=0.5
django-reversion>=3.0.4
django-admin-hstore-widget==1.0.1
PyGithub>=1.44
django-guardian>=2.1.0